+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Pending changes
+---------------
+
+*   Added lenses `ixR`, `ixL`, `ixRow`, and `ixCol`, meant to be used with
+    `^^.` and `viewVar`.
+
 Version 0.1.0.0
 ---------------